+++ app/src/main/java/com/example/zanna/ghioca/MainActivity.java
@@ -1,13 +1,10 @@
 package com.example.zanna.ghioca;
 
-import android.app.AlertDialog;
-import android.app.ProgressDialog;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.AsyncTask;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
-import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.ImageButton;
@@ -16,17 +13,12 @@ import com.flurgle.camerakit.CameraKit;
 import com.flurgle.camerakit.CameraListener;
 import com.flurgle.camerakit.CameraView;
 
-import org.json.JSONException;
-import org.json.JSONObject;
-
 import java.io.File;
-import java.io.FileInputStream;
 import java.sql.Timestamp;
 
 import io.filepicker.Filepicker;
 
 import static com.example.zanna.ghioca.MyApplication.MY_API_KEY;
-import static com.example.zanna.ghioca.SavingUtility.folderPath;
 
 public class MainActivity extends AppCompatActivity {
     final int SELECT_PICTURE = 100;
@@ -96,82 +88,27 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void onPictureTaken(final byte[] jpeg) {
                 super.onPictureTaken(jpeg);
-                new AsyncTask<Void, Void, Void>(){
-
-                    ProgressDialog uploadProgressDialog;
-                    AlertDialog resultDialog;
-                    SearchingListener searchingListener;
-                    UploadingListener uploadingListener;
-
-                    @Override
-                    protected void onPreExecute() {
-                        uploadProgressDialog = new ProgressDialog(MainActivity.this,
-                                ProgressDialog.STYLE_HORIZONTAL);
-                        uploadProgressDialog.setCancelable(false);
-                        uploadProgressDialog.setTitle("Uploading the image");
-                        uploadProgressDialog.show();
-
-                        resultDialog = new AlertDialog.Builder(MainActivity.this).create();
-                        resultDialog.setTitle("Result");
-
-
-                        uploadingListener = new UploadingListener() {
-                            @Override
-                            public void onProgressUpdate(int progress) {
-                                uploadProgressDialog.setProgress(progress);
-                            }
-
-                            @Override
-                            public void onFinish(String url) {
-                                uploadProgressDialog.dismiss();
-                                SearchingUtility.searchImage(url, searchingListener);
-                            }
-
-                            @Override
-                            public void onFailure(Throwable error) {
-                                error.printStackTrace();
-                                uploadProgressDialog.dismiss();
-                            }
-                        };
-
-                        searchingListener = new SearchingListener() {
-                            @Override
-                            public void onFailure(Throwable error) {
-                                error.printStackTrace();
-                                resultDialog.dismiss();
-                            }
-
-                            @Override
-                            public void onSuccess(JSONObject answer) {
-                                try {
-                                    resultDialog.setMessage(answer.getString("best_guess"));
-                                    resultDialog.show();
-                                } catch (JSONException e) {
-                                    e.printStackTrace();
-                                }
-                            }
-                        };
-                    }
-
+                name = photoName();
+                new AsyncTask<Void, Void, Void>() {
                     @Override
                     protected Void doInBackground(Void... params) {
-                        Log.i("provaupload", "1");
-                        name = photoName();
-                        Log.i("provaupload", "2");
                         SavingUtility.saveFile(jpeg, name, MainActivity.this);
-                        Log.i("provaupload", "3");
-                        UploadingUtility.uploadToServer("file://" + folderPath
-                                + "/" + name, MainActivity.this, uploadingListener);
-                        Log.i("provaupload", "4");
                         return null;
                     }
+
+                    @Override
+                    protected void onPostExecute(Void aVoid) {
+                        new AsyncUploadAndSave()
+                                .setContextActivity(MainActivity.this)
+                                .execute(SavingUtility.folderPath + File.separator + name, null, null);
+                    }
                 }.execute(null, null, null);
+
+
+
             }
 
-            private String photoName() {
-                Timestamp timestamp = new Timestamp(System.currentTimeMillis());
-                return timestamp.toString().replaceAll(" ", "_") + ".jpeg";
-            }
+
         });
     }
 
@@ -181,89 +118,20 @@ public class MainActivity extends AppCompatActivity {
                 // Get the url from data
                 final Uri selectedImageUri = data.getData();
                 if (null != selectedImageUri) {
-                    Log.i("asdkjvbgfòadgba", "Image Path : " +
-                            selectedImageUri);
-                    new AsyncTask<Void, Void, Void>(){
-
-                        ProgressDialog uploadProgressDialog;
-                        AlertDialog resultDialog;
-                        SearchingListener searchingListener;
-                        UploadingListener uploadingListener;
-
-                        @Override
-                        protected void onPreExecute() {
-                            uploadProgressDialog = new ProgressDialog(MainActivity.this,
-                                    ProgressDialog.STYLE_HORIZONTAL);
-                            uploadProgressDialog.setCancelable(false);
-                            uploadProgressDialog.setTitle("Uploading the image");
-                            uploadProgressDialog.show();
-
-                            resultDialog = new AlertDialog.Builder(MainActivity.this).create();
-                            resultDialog.setTitle("Result");
-
-
-                            uploadingListener = new UploadingListener() {
-                                @Override
-                                public void onProgressUpdate(int progress) {
-                                    uploadProgressDialog.setProgress(progress);
-                                }
-
-                                @Override
-                                public void onFinish(String url) {
-                                    uploadProgressDialog.dismiss();
-                                    SearchingUtility.searchImage(url, searchingListener);
-                                }
-
-                                @Override
-                                public void onFailure(Throwable error) {
-                                    error.printStackTrace();
-                                    uploadProgressDialog.dismiss();
-                                }
-                            };
-
-                            searchingListener = new SearchingListener() {
-                                @Override
-                                public void onFailure(Throwable error) {
-                                    error.printStackTrace();
-                                    resultDialog.dismiss();
-                                }
-
-                                @Override
-                                public void onSuccess(JSONObject answer) {
-                                    try {
-                                        resultDialog.setMessage(answer.getString("best_guess"));
-                                        resultDialog.show();
-                                    } catch (JSONException e) {
-                                        e.printStackTrace();
-                                    }
-                                }
-                            };
-                        }
-
-                        @Override
-                        protected Void doInBackground(Void... params) {
-                            try {
-                                String filePath = ConvertUriToFilePath.getPathFromURI(MainActivity.this, selectedImageUri);
-                                File file = new File(filePath);
-                                byte[] jpeg = new byte[(int) file.length()];
-
-                                FileInputStream fis = new FileInputStream(file);
-                                fis.read(jpeg); //read file into bytes[]
-                                fis.close();
-                                UploadingUtility.uploadToServer("file://" + filePath,
-                                        MainActivity.this, uploadingListener);
-                            } catch (Exception e) {
-                                e.printStackTrace();
-                            }
-                            return null;
-                        }
-                    }.execute(null, null, null);
+                    new AsyncUploadAndSave()
+                            .setContextActivity(MainActivity.this)
+                            .execute(ConvertUriToFilePath.getPathFromURI(MainActivity.this,
+                                    selectedImageUri), null, null);
                 }
 
             }
         }
     }
 
+    private String photoName() {
+        Timestamp timestamp = new Timestamp(System.currentTimeMillis());
+        return timestamp.toString().replaceAll(" ", "_") + ".jpeg";
+    }
 
     @Override
     protected void onDestroy() {
